Move some impls to a more appropriate place
